+++ src/Server/TaskHandlerChain.php
@@ -4,6 +4,7 @@
 namespace Ingenerator\CloudTasksWrapper\Server;
 
 
+use Ingenerator\CloudTasksWrapper\Server\Middleware\ExceptionCatchingMiddleware;
 use Ingenerator\CloudTasksWrapper\Server\Middleware\TaskLoggingMiddleware;
 use Ingenerator\CloudTasksWrapper\Server\Middleware\TaskMutexLockingMiddleware;
 use Ingenerator\CloudTasksWrapper\Server\Middleware\TaskRequestAuthenticatingMiddleware;
@@ -26,10 +27,12 @@ class TaskHandlerChain
      * The sequence is:
      *
      *  - Logging first so that everything bar dependency issues gets logged
+     *  - Exception catching next to get as many exceptions as possible
      *  - Auth next so that spam calls can't DDOS through the mutex implementation
      *  - Then mutex so that a given task is only processed once at a time
      *
      * @param \Ingenerator\CloudTasksWrapper\Server\Middleware\TaskLoggingMiddleware               $logging
+     * @param \Ingenerator\CloudTasksWrapper\Server\Middleware\ExceptionCatchingMiddleware         $catcher
      * @param \Ingenerator\CloudTasksWrapper\Server\Middleware\TaskRequestAuthenticatingMiddleware $auth
      * @param \Ingenerator\CloudTasksWrapper\Server\Middleware\TaskMutexLockingMiddleware          $mutex
      *
@@ -37,10 +40,11 @@ class TaskHandlerChain
      */
     public static function makeDefault(
         TaskLoggingMiddleware $logging,
+        ExceptionCatchingMiddleware $catcher,
         TaskRequestAuthenticatingMiddleware $auth,
         TaskMutexLockingMiddleware $mutex
     ): TaskHandlerChain {
-        return new static($logging, $auth, $mutex);
+        return new static($logging, $catcher, $auth, $mutex);
     }
 
     public function __construct(TaskHandlerMiddleware...$middlewares)